The scene format supported by SceneLoader was changed to glTF, and corresponding adjustments were made to ASManager

+++ Physically_Based_Rendering_Ray_Tracer/util/Threadsafe_queue.h
@@ -42,7 +42,7 @@ namespace pstd {
 			std::lock_guard<std::mutex> lock(mtx);
 			return data_queue.empty();
 		}
-		unsigned size() {
+		size_t size() {
 			std::lock_guard<std::mutex> lock(mtx);
 			return data_queue.size();
 		}
